+++ docs/changelog.md
@@ -14,6 +14,7 @@ Versioning currently follows `X.Y.Z` where
 
 - Add `Workunit`, `Parameter`, and `Resource` entities.
 - Add concept of has_many and has_one relationships to entities.
+- `bfabric_slurm_queue_status.py` to quickly check slurm queue status.
 
 ## \[1.13.3\] - 2024-07-18